pyproject.toml
# License: MIT
# Copyright © 2024 Frequenz Energy-as-a-Service GmbH

[build-system]
requires = [
  "setuptools == 68.1.0",
  "setuptools_scm[toml] == 7.1.0",
  "frequenz-repo-config[lib] == 0.9.2",
]
build-backend = "setuptools.build_meta"

[project]
name = "frequenz-client-reporting"
description = "Reporting API client for Python"
readme = "README.md"
license = { text = "MIT" }
keywords = ["frequenz", "python", "lib", "library", "client-reporting", "client", "reporting", "api"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Topic :: Software Development :: Libraries",
  "Typing :: Typed",
]
requires-python = ">= 3.11, < 4"
dependencies = [
  "typing_extensions >= 4.6.1, < 5",
  "grpc-stubs == 1.53.0.5",
  "frequenz-api-reporting >= 0.2.0, < 0.4",
  "frequenz-client-common >= 0.2.0, < 0.3",
  "grpcio >= 1.54.2, < 2",
  "grpcio-tools >= 1.54.2, < 2",
  "protobuf >= 4.25.3, < 6",
  "frequenz-client-base[grpcio] >= 0.5.0, < 0.6.0",
]
dynamic = ["version"]

[project.scripts]
reporting-cli = "frequenz.client.reporting.__main__:main"

[[project.authors]]
name = "Frequenz Energy-as-a-Service GmbH"
email = "floss@frequenz.com"

[project.optional-dependencies]
dev-flake8 = [
  "flake8 == 7.1.0",
  "flake8-docstrings == 1.7.0",
  "flake8-pyproject == 1.2.3",  # For reading the flake8 config from pyproject.toml
  "pydoclint == 0.4.1",
  "pydocstyle == 6.3.0",
]
dev-formatting = ["black == 24.4.2", "isort == 5.13.2"]
dev-mkdocs = [
  "black == 24.4.2",
  "Markdown==3.6",
  "mike == 2.1.1",
  "mkdocs-gen-files == 0.5.0",
  "mkdocs-literate-nav == 0.6.1",
  "mkdocs-macros-plugin == 1.0.5",
  "mkdocs-material == 9.5.25",
  "mkdocstrings[python] == 0.25.1",
  "frequenz-repo-config[lib] == 0.9.2",
]
dev-mypy = [
  "mypy == 1.10.0",
  "types-Markdown == 3.6.0.20240316",
  "types-protobuf == 5.26.0.20240422",
  # For checking the noxfile, docs/ script, and tests
  "frequenz-client-reporting[dev-mkdocs,dev-noxfile,dev-pytest]",
  "pandas-stubs >= 2, < 3",  # Only required for example
]
dev-noxfile = [
  "nox == 2024.4.15",
  "frequenz-repo-config[lib] == 0.9.2",
]
dev-pylint = [
  "pylint == 3.2.2",
  # For checking the noxfile, docs/ script, and tests
  "frequenz-client-reporting[dev-mkdocs,dev-noxfile,dev-pytest]",
  "pandas >= 2, < 3",  # Only required for example
]
dev-pytest = [
  "pytest == 8.2.1",
  "frequenz-repo-config[extra-lint-examples] == 0.9.2",
  "pytest-mock == 3.14.0",
  "pytest-asyncio == 0.23.7",
  "async-solipsism == 0.6",
]
dev = [
  "frequenz-client-reporting[dev-mkdocs,dev-flake8,dev-formatting,dev-mkdocs,dev-mypy,dev-noxfile,dev-pylint,dev-pytest]",
]
examples = [
    "pandas >= 2, < 3",
]

[project.urls]
Documentation = "https://frequenz-floss.github.io/frequenz-client-reporting-python/"
Changelog = "https://github.com/frequenz-floss/frequenz-client-reporting-python/releases"
Issues = "https://github.com/frequenz-floss/frequenz-client-reporting-python/issues"
Repository = "https://github.com/frequenz-floss/frequenz-client-reporting-python"
Support = "https://github.com/frequenz-floss/frequenz-client-reporting-python/discussions/categories/support"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
src_paths = ["benchmarks", "examples", "src", "tests"]

[tool.flake8]
# We give some flexibility to go over 88, there are cases like long URLs or
# code in documenation that have extra indentation. Black will still take care
# of making everything that can be 88 wide, 88 wide.
max-line-length = 100
extend-ignore = [
  "E203", # Whitespace before ':' (conflicts with black)
  "W503", # Line break before binary operator (conflicts with black)
]
# pydoclint options
style = "google"
check-return-types = false
check-yield-types = false
arg-type-hints-in-docstring = false
arg-type-hints-in-signature = true
allow-init-docstring = true

[tool.pylint.similarities]
ignore-comments = ['yes']
ignore-docstrings = ['yes']
ignore-imports = ['no']
min-similarity-lines = 40

[tool.pylint.messages_control]
disable = [
  "too-few-public-methods",
  "too-many-return-statements",
  # disabled because it conflicts with isort
  "wrong-import-order",
  "ungrouped-imports",
  # pylint's unsubscriptable check is buggy and is not needed because
  # it is a type-check, for which we already have mypy.
  "unsubscriptable-object",
  # Checked by flake8
  "line-too-long",
  "redefined-outer-name",
  "unnecessary-lambda-assignment",
  "unused-import",
  "unused-variable",
]

[tool.pytest.ini_options]
testpaths = ["tests", "src"]
asyncio_mode = "auto"
required_plugins = ["pytest-asyncio", "pytest-mock"]

[tool.mypy]
explicit_package_bases = true
namespace_packages = true
# This option disables mypy cache, and it is sometimes useful to enable it if
# you are getting weird intermittent error, or error in the CI but not locally
# (or vice versa). In particular errors saying that type: ignore is not
# used but getting the original ignored error when removing the type: ignore.
# See for example: https://github.com/python/mypy/issues/2960
no_incremental = true
packages = ["frequenz.client.reporting"]
strict = true

[[tool.mypy.overrides]]
module = ["grpc.aio", "grpc.aio.*", "mkdocs_macros.*", "sybil", "sybil.*"]
ignore_missing_imports = true

[tool.setuptools_scm]
version_scheme = "post-release"
